requirements.txt
Django==1.6.2
South==0.8.4
dj-database-url==0.3.0
dj-static==0.0.5
django-annoying==0.7.9
django-toolbelt==0.0.1
docopt==0.4.0
gunicorn==18.0
heroku==0.1.4
mandrill==1.0.53
psycopg2==2.5.2
pystache==0.5.3
python-dateutil==1.5
requests==2.2.1
static==1.0.2
tweepy==2.2
wsgiref==0.1.2
